bug fix, to ensure Solutions obj array is changed in-place in SolutionSample

+++ python3/problem1/shd_1_plus_lambda.py
@@ -35,7 +35,7 @@ from problem1 import Printer3d
 p = Printer3d(200)
 
 imax = 2000 # max iterations
-out = np.zeros(imax, int)
+out = np.zeros(imax, float)
 n = 5 # no. of children
 x = p.randomSolution(n+1)
 fx = x.objvalue()
